Create armorwearevents.js
quick scope on how i think kubejs should work 
it is translating from java to kubejs so i dont know if this is well defined.
It should allow you to check whenever a player takes on or wears off an armor piece. Runs each tick so we gotta optimize (reason why I used a weak Hashmap.